requirements.txt
flask
flask-cors

# GitHub Copilot SDK Demo - Requirements
github-copilot-sdk
pydantic
